requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
pytest==7.4.3
pytest-django==4.7.0
flake8==6.1.0
black==23.11.0
isort==5.12.0
